REF: Use react-native-screen-capture
