+++ package.json
@@ -138,15 +138,12 @@
     "react-native-gesture-handler": "2.20.0",
     "react-native-handoff": "github:BlueWallet/react-native-handoff#v0.0.4",
     "react-native-haptic-feedback": "2.3.3",
-    "react-native-idle-timer": "github:BlueWallet/react-native-idle-timer#v2.2.3",
     "react-native-image-picker": "7.1.2",
     "react-native-ios-context-menu": "github:BlueWallet/react-native-ios-context-menu#e5c1217cd220bfab6e6d9a7c65838545082e3f8e",
     "react-native-keychain": "8.2.0",
     "react-native-linear-gradient": "2.8.3",
     "react-native-localize": "3.2.1",
-    "react-native-obscure": "github:BlueWallet/react-native-obscure#f4b83b4a261e39b1f5ed4a45ac5bcabc8a59eadb",
     "react-native-permissions": "4.1.5",
-    "react-native-privacy-snapshot": "github:BlueWallet/react-native-privacy-snapshot#529e4627d93f67752a27e82a040ff7b64dca0783",
     "react-native-prompt-android": "github:BlueWallet/react-native-prompt-android#ed168d66fed556bc2ed07cf498770f058b78a376",
     "react-native-push-notification": "8.1.1",
     "react-native-qrcode-svg": "6.3.2",
@@ -155,6 +152,7 @@
     "react-native-rate": "1.2.12",
     "react-native-reanimated": "3.15.5",
     "react-native-safe-area-context": "4.11.0",
+    "react-native-screen-capture": "github:BlueWallet/react-native-screen-capture#18cb79f",
     "react-native-screens": "3.34.0",
     "react-native-secure-key-store": "github:BlueWallet/react-native-secure-key-store#2076b4849e88aa0a78e08bfbb4ce3923e0925cbc",
     "react-native-share": "10.2.1",
