+++ tests/setup.js
@@ -159,12 +159,6 @@ jest.mock('realm', () => {
   };
 });
 
-jest.mock('react-native-idle-timer', () => {
-  return {
-    setIdleTimerDisabled: jest.fn(),
-  };
-});
-
 jest.mock('react-native-ios-context-menu', () => {
   return {};
 });
